+++ lib/influxdb/query/retention_policy.rb
@@ -1,10 +1,14 @@
 module InfluxDB
   module Query
     module RetentionPolicy # :nodoc:
-      def create_retention_policy(name, database, duration, replication, default = false)
+      def create_retention_policy(name, database, duration, replication, options = {})
+        defaults = { default: false, shard_duration: nil }
+        options = defaults.merge(options)
         execute(
           "CREATE RETENTION POLICY \"#{name}\" ON #{database} " \
-          "DURATION #{duration} REPLICATION #{replication}#{default ? ' DEFAULT' : ''}"
+          "DURATION #{duration} REPLICATION #{replication}" \
+          "#{options[:shard_duration] ? " SHARD DURATION #{options[:shard_duration]}" : ''}" \
+          "#{options[:default] ? ' DEFAULT' : ''}"
         )
       end
 
@@ -23,10 +27,14 @@ module InfluxDB
         execute("DROP RETENTION POLICY \"#{name}\" ON #{database}")
       end
 
-      def alter_retention_policy(name, database, duration, replication, default = false)
+      def alter_retention_policy(name, database, duration, replication, options = {})
+        defaults = { default: false, shard_duration: nil }
+        options = defaults.merge(options)
         execute(
           "ALTER RETENTION POLICY \"#{name}\" ON #{database} " \
-          "DURATION #{duration} REPLICATION #{replication}#{default ? ' DEFAULT' : ''}"
+          "DURATION #{duration} REPLICATION #{replication}" \
+          "#{options[:shard_duration] ? " SHARD DURATION #{options[:shard_duration]}" : ''}" \
+          "#{options[:default] ? ' DEFAULT' : ''}"
         )
       end
     end