+++ spec/influxdb/cases/query_retention_policy_spec.rb
@@ -40,7 +40,7 @@ describe InfluxDB::Client do
       let(:query) { "CREATE RETENTION POLICY \"1h.cpu\" ON foo DURATION 1h REPLICATION 2 DEFAULT" }
 
       it "should GET to create a new database" do
-        expect(subject.create_retention_policy('1h.cpu', 'foo', '1h', 2, true)).to be_a(Net::HTTPOK)
+        expect(subject.create_retention_policy('1h.cpu', 'foo', '1h', 2, default: true)).to be_a(Net::HTTPOK)
       end
     end
 
@@ -49,6 +49,22 @@ describe InfluxDB::Client do
 
       it "should GET to create a new database" do
         expect(subject.create_retention_policy('1h.cpu', 'foo', '1h', 2)).to be_a(Net::HTTPOK)
+      end
+    end
+
+    context "default_with_shard_duration" do
+      let(:query) { "CREATE RETENTION POLICY \"1h.cpu\" ON foo DURATION 48h REPLICATION 2 SHARD DURATION 1h DEFAULT" }
+
+      it "should GET to create a new database" do
+        expect(subject.create_retention_policy('1h.cpu', 'foo', '48h', 2, default: true, shard_duration: '1h')).to be_a(Net::HTTPOK)
+      end
+    end
+
+    context "non-default_with_shard_duration" do
+      let(:query) { "CREATE RETENTION POLICY \"1h.cpu\" ON foo DURATION 48h REPLICATION 2 SHARD DURATION 1h" }
+
+      it "should GET to create a new database" do
+        expect(subject.create_retention_policy('1h.cpu', 'foo', '48h', 2, shard_duration: '1h')).to be_a(Net::HTTPOK)
       end
     end
   end
@@ -66,7 +82,7 @@ describe InfluxDB::Client do
       let(:query) { "ALTER RETENTION POLICY \"1h.cpu\" ON foo DURATION 1h REPLICATION 2 DEFAULT" }
 
       it "should GET to alter a new database" do
-        expect(subject.alter_retention_policy('1h.cpu', 'foo', '1h', 2, true)).to be_a(Net::HTTPOK)
+        expect(subject.alter_retention_policy('1h.cpu', 'foo', '1h', 2, default: true)).to be_a(Net::HTTPOK)
       end
     end
 
@@ -75,6 +91,22 @@ describe InfluxDB::Client do
 
       it "should GET to alter a new database" do
         expect(subject.alter_retention_policy('1h.cpu', 'foo', '1h', 2)).to be_a(Net::HTTPOK)
+      end
+    end
+
+    context "default_with_shard_duration" do
+      let(:query) { "ALTER RETENTION POLICY \"1h.cpu\" ON foo DURATION 48h REPLICATION 2 SHARD DURATION 1h DEFAULT" }
+
+      it "should GET to alter a new database" do
+        expect(subject.alter_retention_policy('1h.cpu', 'foo', '48h', 2, default: true, shard_duration: '1h')).to be_a(Net::HTTPOK)
+      end
+    end
+
+    context "non-default_with_shard_duration" do
+      let(:query) { "ALTER RETENTION POLICY \"1h.cpu\" ON foo DURATION 48h REPLICATION 2 SHARD DURATION 1h" }
+
+      it "should GET to alter a new database" do
+        expect(subject.alter_retention_policy('1h.cpu', 'foo', '48h', 2, shard_duration: '1h')).to be_a(Net::HTTPOK)
       end
     end
   end